linux_i2c: Set the iBus as environment variable
One of the possible problems folks face is when multiple i2c bus is
used for various functions, in such cases, the dt_overlay mechanism to
setup the board to map the desired i2cbus to i2c1 may not be available
for various reasons.

See [1][2][3] as examples. Lets just skip this problem by allowing
people to quickly prototype solutions by exporting the I2C bus -
though this solution will not nicely scale if there are additional
qwiic connectors on the board which need to be controlled by the same
application script.

[1] https://forum.sparkfun.com/viewtopic.php?f=105&t=54688
[2] https://github.com/sparkfun/Qwiic_I2C_Py/pull/1
[3] https://github.com/sparkfun/Qwiic_I2C_Py/pull/8

Signed-off-by: Nishanth Menon <nm@ti.com>