Add proper ping method
Previous implementation wrote a zero byte to the device. Works fine most of the time, but turns out not everywhere! Discovered the CAP1203 gets into a weird state if you do this, for example you can't read the product ID register properly.

This adds a dedicated isDeviceConnected() method to ever platform driver, so no extra bytes are written. Seems to be more happy now!